Potential fix of date time utility for Eu time zone date format

+++ Slack.Notification.Desktop.Application/MainWindow.xaml.cs
@@ -173,7 +173,7 @@ namespace SlackDesktopBubbleApplication
 
             foreach (var element in elements)
             {
-                double messageTimestamp = Convert.ToDouble(element.Uid); //TODO: CAN NOT CONVERT 1481889334.00001
+                double messageTimestamp = Convert.ToInt32(element.Uid); //TODO: CAN NOT CONVERT 1481889334.00001
 
                 var messageAddedAt = DateTimeUtility.UnixTimeStampToDateTime(messageTimestamp);
